+++ internal/storage/authn/auth.go
@@ -1,4 +1,4 @@
-package auth
+package authn
 
 import (
 	"bytes"
