+++ internal/storage/authn/auth_test.go
@@ -1,4 +1,4 @@
-package auth
+package authn
 
 import (
 	"encoding/base64"
