+++ internal/storage/authn/bootstrap.go
@@ -1,4 +1,4 @@
-package auth
+package authn
 
 import (
 	"context"
