+++ internal/storage/authn/sql/store_test.go
@@ -10,8 +10,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.flipt.io/flipt/errors"
 	"go.flipt.io/flipt/internal/storage"
-	storageauth "go.flipt.io/flipt/internal/storage/auth"
-	authtesting "go.flipt.io/flipt/internal/storage/auth/testing"
+	storageauth "go.flipt.io/flipt/internal/storage/authn"
+	authtesting "go.flipt.io/flipt/internal/storage/authn/testing"
 	storagesql "go.flipt.io/flipt/internal/storage/sql"
 	sqltesting "go.flipt.io/flipt/internal/storage/sql/testing"
 	rpcauth "go.flipt.io/flipt/rpc/flipt/auth"
